feat(lambda): move DI factory interface into infrastructure layer #59
